feat(dev): expose restart function on ViteDevServer
