+++ packages/vite/src/node/server/hmr.ts
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import path from 'path'
 import chalk from 'chalk'
-import { createServer, ViteDevServer } from '..'
+import { ViteDevServer } from '..'
 import { createDebugger, normalizePath } from '../utils'
 import { ModuleNode } from './moduleGraph'
 import { Update } from 'types/hmrPayload'
@@ -10,7 +10,6 @@ import { RollupError } from 'rollup'
 import { isMatch } from 'micromatch'
 import { Server } from 'http'
 import { isCSSRequest } from '../plugins/css'
-import { performance } from 'perf_hooks'
 
 export const debugHmr = createDebugger('vite:hmr')
 
@@ -62,7 +61,7 @@ export async function handleHMRUpdate(
       ),
       { clear: true, timestamp: true }
     )
-    await restartServer(server)
+    await server.restart()
     return
   }
 
@@ -460,35 +459,5 @@ async function readModifiedFile(file: string): Promise<string> {
     return fs.readFileSync(file, 'utf-8')
   } else {
     return content
-  }
-}
-
-async function restartServer(server: ViteDevServer) {
-  // @ts-ignore
-  global.__vite_start_time = performance.now()
-  const { port } = server.config.server
-
-  await server.close()
-
-  let newServer = null
-  try {
-    newServer = await createServer(server.config.inlineConfig)
-  } catch (err: any) {
-    server.config.logger.error(err.message, {
-      timestamp: true
-    })
-    return
-  }
-
-  for (const key in newServer) {
-    if (key !== 'app') {
-      // @ts-ignore
-      server[key] = newServer[key]
-    }
-  }
-  if (!server.config.server.middlewareMode) {
-    await server.listen(port, true)
-  } else {
-    server.config.logger.info('server restarted.', { timestamp: true })
   }
 }
